Bug 1054947 - Fix misjudging user who always uses and never closes Firefox as one that hasn't used Firefox for a long time
This patch switches to using the last modified time of the profiles.ini file to check the last Firefox-used time. Also update the last modified time of the profiles.ini for every 24 hours so that it wouldn't misjudge the case that user hadn't used Firefox for a long time if user never closed Firefox for months.

MozReview-Commit-ID: K1Rh3MQ7Kxv