Bug 1370780 - Test storage pressure notification according to the browser.preferences.useOldOrganization pref
On Nightly it is the new about:preferences (browser.preferences.useOldOrganization = true) so the siteDataGroup is under about:preferences#advanced.
On Beta and Release it is the old about:preferences (browser.preferences.useOldOrganization = false) so the siteDataGroup is under about:preferences#privacy.

This patch will detect browser.preferences.useOldOrganization to decide the right about:preferences url to check

MozReview-Commit-ID: CFm0ntHfRq6